add fix hw

+++ src/controllers/notesController.js
@@ -8,7 +8,7 @@ export const getAllNotes = async (req, res) => {
 
   req.user;
 
-  const notesQuery = Note.find();
+  const notesQuery = Note.find({userId: req.user._id});
 
   if (search) {
     notesQuery.where({
@@ -40,7 +40,7 @@ export const getAllNotes = async (req, res) => {
 
 export const getNoteById = async (req, res) => {
   const { noteId } = req.params;
-  const note = await Note.findOne({ _id: noteId });
+  const note = await Note.findOne({ _id: noteId, userId: req.user._id });
 
   if (!note) {
     // return res.status(404).json({message: "❌ Note not found"});
@@ -51,14 +51,17 @@ export const getNoteById = async (req, res) => {
 
 export const createNote = async (req, res) => {
   console.log('CONTROLLER -> createNote');
-  const note = await Note.create(req.body);
+  const note = await Note.create({
+    ...req.body,
+    userId:req.user._id,
+  });
   res.status(201).json(note);
   // res.status(201).json(note);
 };
 
 export const deleteNote = async (req, res) => {
   const { noteId } = req.params;
-  const note = await Note.findOneAndDelete({ _id: noteId });
+  const note = await Note.findOneAndDelete({ _id: noteId , userId: req.user._id});
   if (!note) {
     throw createHttpError(404, '❌ Note not found');
   }
@@ -67,7 +70,7 @@ export const deleteNote = async (req, res) => {
 
 export const updateNote = async (req, res) => {
   const { noteId } = req.params;
-  const note = await Note.findOneAndUpdate({ _id: noteId }, req.body, {
+  const note = await Note.findOneAndUpdate({ _id: noteId, userId: req.user._id }, req.body, {
     returnDocument: 'after',
   });
   if (!note) {
